+++ source/ApplyDBMigrationsApp/Extensibility/DbUp/ScriptComparer.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.EDI.ApplyDBMigrationsApp.Helpers;
+namespace Energinet.DataHub.EDI.ApplyDBMigrationsApp.Extensibility.DbUp;
 
 /// <summary>
 /// Compare sql script file names
