+++ source/ApplyDBMigrationsApp/Scripts/Permissions/202512061200 Grant access to query execution plan.sql
